+++ Lib/fontbakery/specifications/googlefonts.py
@@ -115,7 +115,7 @@ expected_check_ids = \
       , 'com.google.fonts/check/unitsperem'
       , 'com.google.fonts/check/dsig'
       , 'com.google.fonts/check/mandatory_glyphs'
-      , 'com.google.fonts/check/047' # Font contains glyphs for whitespace characters?
+      , 'com.google.fonts/check/whitespace_glyphs'
       , 'com.google.fonts/check/048' # Font has **proper** whitespace glyph names?
       , 'com.google.fonts/check/049' # Whitespace glyphs have ink?
       , 'com.google.fonts/check/050' # Whitespace glyphs have coherent widths?